refactor(walletconnect): extract shared WC pure helpers to @safe-global/utils
Move the dependency-free WalletConnect helpers into a new
packages/utils/src/features/walletconnect module so web and mobile share
one implementation instead of duplicating:

- EIP-155/CAIP + URI + error helpers (getEip155ChainId, stripEip155Prefix,
  chainIdToHex, isPairingUri, splitError)
- EIP-5792 atomic-batch capability builder (buildAtomicCapabilities)
- verify-status -> UI variant mapping (decoupled from @walletconnect/types
  via a structural VerifiedContext, keeping the package dependency-free)

Web re-exports the four pure helpers from the shared module; mobile services
consume them directly, dropping hand-rolled prefix/hex math and the local
capability map. The existing splitError test moves alongside its subject.
