+++ apps/mobile/src/features/WalletConnect/Wallet/utils/verifyStatus.ts
@@ -1,16 +1,1 @@
-import type { Verify } from '@walletconnect/types'
-
-export type VerifyVariant = 'verified' | 'unverified' | 'malicious'
-
-export const verifyStatusToVariant = (verifyContext?: Verify.Context['verified']): VerifyVariant => {
-  if (!verifyContext) {
-    return 'unverified'
-  }
-  if (verifyContext.isScam) {
-    return 'malicious'
-  }
-  if (verifyContext.validation === 'VALID') {
-    return 'verified'
-  }
-  return 'unverified' // UNKNOWN or INVALID
-}
+export { verifyStatusToVariant, type VerifyVariant } from '@safe-global/utils/features/walletconnect/verify'
