+++ packages/utils/src/features/walletconnect/__tests__/utils.test.ts
