+++ packages/utils/src/features/walletconnect/constants.ts
@@ -1,0 +1,2 @@
+// The only WalletConnect namespace Safe supports.
+export const EIP155 = 'eip155' as const
